// src/app/modules/support/models/faq.model.ts
export interface IFaqModel {
  updateDate: string;
  typeNameAz: string;
  createUserId: string;
  question: string;
  updateUserId: string;
  typeName: string;
  active: string;
  questionRu: string;
  searchKeyword: string;
  questionEn: string;
  answerAz: string;
  typeNameEn: string;
  answerFile: string;
  answer: string;
  typeNameRu: string;
  questionFile: string;
  answerRu: string;
  typeId: string;
  userType: string;
  id: string;
  questionAz: string;
  answerEn: string;
  createDate: string;
}